+++ backend/src/error.rs
@@ -5,7 +5,7 @@ use reqwest::StatusCode;
 use shared::{AddQuestionValidation, PasswordValidation, TagValidation};
 use thiserror::Error;
 
-use crate::{eventsdb, payment::PaymentError, plots::PlottingError};
+use crate::{eventsdb, payment::PaymentError, tracking};
 
 #[derive(Error, Debug)]
 pub enum InternalError {
@@ -60,8 +60,8 @@ pub enum InternalError {
     #[error("Uri Error: {0}")]
     Uri(#[from] axum::http::uri::InvalidUri),
 
-    #[error("Plotting Error: {0}")]
-    PlottingError(#[from] PlottingError),
+    #[error("Tracking Error: {0}")]
+    Tracking(#[from] tracking::TrackingError),
 }
 
 impl IntoResponse for InternalError {
@@ -69,11 +69,6 @@ impl IntoResponse for InternalError {
     fn into_response(self) -> Response {
         match self {
             Self::General(e) => {
-                tracing::error!("{e}");
-                (StatusCode::INTERNAL_SERVER_ERROR, "").into_response()
-            }
-
-            Self::PlottingError(e) => {
                 tracing::error!("{e}");
                 (StatusCode::INTERNAL_SERVER_ERROR, "").into_response()
             }
@@ -151,6 +146,7 @@ impl IntoResponse for InternalError {
             Self::Uri(e) => convert_error(e),
             Self::DeadPoolCreatePool(e) => convert_error(e),
             Self::DeadPoolRedis(e) => convert_error(e),
+            Self::Tracking(e) => convert_error(e),
         }
     }
 }