+++ backend/src/payment/mod.rs
@@ -36,7 +36,7 @@ impl Payment {
         }
     }
 
-    pub async fn authenticate(&mut self) -> PaymentResult<bool> {
+    pub async fn authenticate(&mut self, live: bool) -> PaymentResult<String> {
         let res = stripe::Product::list(&self.client, &ListProducts::new()).await?;
 
         for p in &res.data {
@@ -52,25 +52,27 @@ impl Payment {
             if p.metadata
                 .as_ref()
                 .and_then(|meta| meta.get("id"))
-                .map(|id| id == "premium")
-                .unwrap_or_default()
+                .is_some_and(|id| id == "premium")
             {
                 tracing::info!("[stripe] prod id: {:?} is premium package", p.id);
-                self.price = Some(
-                    p.default_price
-                        .as_ref()
-                        .ok_or_else(|| PaymentError::Generic("default price not set".into()))?
-                        .id()
-                        .to_string(),
-                );
 
-                if p.livemode.unwrap_or_default() {
-                    return Ok(true);
+                if p.livemode.unwrap_or_default() == live {
+                    self.price = Some(
+                        p.default_price
+                            .as_ref()
+                            .ok_or_else(|| PaymentError::Generic("default price not set".into()))?
+                            .id()
+                            .to_string(),
+                    );
+
+                    return Ok(p.id.to_string());
                 }
             }
         }
 
-        Ok(false)
+        Err(PaymentError::Generic(String::from(
+            "no premium product found",
+        )))
     }
 
     pub async fn create_order(
@@ -84,10 +86,14 @@ impl Payment {
             params.cancel_url = Some(mod_url);
             params.success_url = Some(return_url);
             params.client_reference_id = Some(event);
+            params.allow_promotion_codes = Some(true);
             params.metadata = Some(
-                vec![(String::from("event"), event.to_string())]
-                    .into_iter()
-                    .collect(),
+                vec![
+                    (String::from("event"), event.to_string()),
+                    (String::from("url"), mod_url.to_string()),
+                ]
+                .into_iter()
+                .collect(),
             );
             params.mode = Some(CheckoutSessionMode::Payment);
             params.line_items = Some(vec![CreateCheckoutSessionLineItems {
@@ -116,8 +122,7 @@ impl Payment {
         let event = sess.client_reference_id.unwrap_or_default();
         let completed = sess
             .status
-            .map(|status| status == CheckoutSessionStatus::Complete)
-            .unwrap_or_default();
+            .is_some_and(|status| status == CheckoutSessionStatus::Complete);
 
         Ok((event, completed))
     }