+++ frontend/src/pages/event.rs
@@ -2,13 +2,13 @@ use chrono::{DateTime, Duration, NaiveDateTime, Utc};
 use const_format::formatcp;
 use events::{event_context, EventBridge};
 use serde::Deserialize;
-use shared::{EventInfo, GetEventResponse, ModEvent, ModQuestion, QuestionItem, States};
+use shared::{
+    EventFlags, EventInfo, GetEventResponse, ModEvent, ModQuestion, QuestionItem, States,
+};
 use std::{collections::HashMap, rc::Rc, str::FromStr};
 use wasm_bindgen::{JsCast, UnwrapThrowExt};
 use web_sys::HtmlAnchorElement;
-use worker2::{WordCloudAgent, WordCloudInput, WordCloudOutput};
 use yew::{prelude::*, virtual_dom::AttrValue};
-use yew_agent::{Bridge, Bridged};
 use yew_router::scope_ext::RouterScopeExt;
 use yewdux::prelude::*;
 
@@ -69,7 +69,6 @@ struct QueryParams {
 pub struct Event {
     current_event_id: String,
     copied_to_clipboard: bool,
-    wordcloud: Option<String>,
     query_params: QueryParams,
     mode: Mode,
     tags: SharableTags,
@@ -81,7 +80,6 @@ pub struct Event {
     state: Rc<State>,
     dispatch: Dispatch<State>,
     events: EventBridge<GlobalEvent>,
-    wordcloud_agent: Box<dyn Bridge<WordCloudAgent>>,
     socket_url: String,
     manual_reconnect: bool,
 }
@@ -102,7 +100,6 @@ pub enum Msg {
     CopyLink,
     ModEditScreening,
     GlobalEvent(GlobalEvent),
-    WordCloud(WordCloudOutput),
 }
 impl Component for Event {
     type Message = Msg;
@@ -129,17 +126,11 @@ impl Component for Event {
             .unwrap_throw()
             .subscribe(ctx.link().callback(Msg::GlobalEvent));
 
-        let cb_wordcloud = {
-            let link = ctx.link().clone();
-            move |e| link.send_message(Msg::WordCloud(e))
-        };
-
         let dispatch = Dispatch::<State>::subscribe(Callback::noop());
 
         Self {
             current_event_id: event_id,
             query_params,
-            wordcloud: None,
             copied_to_clipboard: false,
             mode: if ctx.props().secret.is_some() {
                 Mode::Moderator
@@ -155,7 +146,6 @@ impl Component for Event {
             unscreened: Vec::new(),
             dispatch,
             events,
-            wordcloud_agent: WordCloudAgent::bridge(Rc::new(cb_wordcloud)),
             socket_url,
             manual_reconnect: false,
         }
@@ -212,8 +202,7 @@ impl Component for Event {
                             self.state
                                 .event
                                 .as_ref()
-                                .map(|e| !e.info.is_screening())
-                                .unwrap_or_default(),
+                                .is_some_and(|e| !e.info.is_screening()),
                         ),
                         ..Default::default()
                     },
@@ -221,19 +210,6 @@ impl Component for Event {
                 );
 
                 false
-            }
-            Msg::WordCloud(w) => {
-                if self
-                    .wordcloud
-                    .as_ref()
-                    .map(|old| old == &w.0)
-                    .unwrap_or_default()
-                {
-                    false
-                } else {
-                    self.wordcloud = Some(w.0);
-                    true
-                }
             }
             Msg::ModDelete => {
                 self.events.emit(GlobalEvent::DeletePopup);
@@ -412,8 +388,7 @@ impl Event {
         self.state
             .event
             .as_ref()
-            .map(|e| e.info.is_premium())
-            .unwrap_or_default()
+            .is_some_and(|e| e.info.is_premium())
     }
 
     fn export_event(&self) {
@@ -521,6 +496,7 @@ impl Event {
             let admin = e.admin;
 
             let tag = e.info.tags.get_current_tag_label();
+            let screening_enabled = e.info.flags.contains(EventFlags::SCREENING);
 
             html! {
                 <div class="some-event">
@@ -557,35 +533,18 @@ impl Event {
 
                     {self.mod_urls(ctx,admin)}
 
-                    {self.view_viewers()}
+                    {self.view_stats()}
+
+                    <div class="review-note" hidden={!screening_enabled || mod_view}>
+                        {"Moderator enabled question reviewing. New questions have to be approved first."}
+                    </div>
 
                     {self.view_questions(ctx,e)}
-
-                    {self.view_cloud(e)}
 
                     {Self::view_ask_question(mod_view,ctx,e)}
                 </div>
             }
         })
-    }
-
-    fn view_cloud(&self, e: &GetEventResponse) -> Html {
-        let title_classes = self.question_separator_classes();
-
-        if let Some(wc) = &self.wordcloud {
-            if !e.masked {
-                return html! {
-                    <div>
-                    <div class={title_classes}>
-                        {"Word Cloud"}
-                    </div>
-                    {cloud_as_yew_img(wc)}
-                    </div>
-                };
-            }
-        }
-
-        html!()
     }
 
     #[allow(clippy::if_not_else)]
@@ -615,7 +574,7 @@ impl Event {
     }
 
     fn view_questions(&self, ctx: &Context<Self>, e: &GetEventResponse) -> Html {
-        if e.info.questions.is_empty() {
+        if e.info.questions.is_empty() && self.unscreened.is_empty() {
             let no_questions_classes = classes!(match self.mode {
                 Mode::Moderator => "noquestions modview",
                 Mode::Viewer => "noquestions",
@@ -650,12 +609,7 @@ impl Event {
         if !items.is_empty() {
             let title_classes = self.question_separator_classes();
 
-            let masked = self
-                .state
-                .event
-                .as_ref()
-                .map(|e| e.masked)
-                .unwrap_or_default();
+            let masked = self.state.event.as_ref().is_some_and(|e| e.masked);
 
             return html! {
                 <div>
@@ -684,11 +638,7 @@ impl Event {
     ) -> Html {
         let local_like = LocalCache::is_liked(&self.current_event_id, item.id);
         let mod_view = matches!(self.mode, Mode::Moderator);
-        let is_new = self
-            .state
-            .new_question
-            .map(|id| id == item.id)
-            .unwrap_or_default();
+        let is_new = self.state.new_question.is_some_and(|id| id == item.id);
 
         let mut flags = QuestionFlags::empty();
 
@@ -800,7 +750,7 @@ impl Event {
         }
     }
 
-    fn view_viewers(&self) -> Html {
+    fn view_stats(&self) -> Html {
         if !self.is_premium() {
             return html! {};
         }
@@ -874,7 +824,7 @@ impl Event {
                         </button>
                         <button class="button-blue">
                             <a class="feedback-anchor"
-                                href="https://forms.gle/DsD9ZEX5uv1QqDjV7"
+                                href="https://noteforms.com/forms/user-feedback-wlfknf"
                                 target="_blank"
                                 onclick={ctx.link().callback(|_| Msg::FeedbackClick)}>
                                 <div class="feedback-text">{"Give us feedback"}</div>
@@ -948,16 +898,6 @@ impl Event {
                 .map(|t| (t.id, t.name.clone()))
                 .collect::<HashMap<_, _>>()
                 .into();
-
-            if e.info.is_premium() && !e.masked && e.any_questions() {
-                self.wordcloud_agent.send(WordCloudInput(
-                    e.info
-                        .questions
-                        .iter()
-                        .map(|q| q.text.clone())
-                        .collect::<Vec<_>>(),
-                ));
-            }
         }
     }
 
@@ -1075,8 +1015,7 @@ impl Event {
                         .state
                         .event
                         .as_ref()
-                        .map(|e| e.info.questions.iter().any(|q| q.id == id))
-                        .unwrap_or_default();
+                        .is_some_and(|e| e.info.questions.iter().any(|q| q.id == id));
 
                     if !found {
                         log::info!("new question: {}", id);
@@ -1140,23 +1079,11 @@ impl Event {
             if !e.info.is_premium() && self.query_params.paypal_token.is_some() {
                 request_capture(
                     e.info.tokens.public_token,
-                    self.query_params
-                        .paypal_token
-                        .as_ref()
-                        .cloned()
-                        .unwrap_throw(),
+                    self.query_params.paypal_token.clone().unwrap_throw(),
                     ctx.link(),
                 );
             }
         }
         true
-    }
-}
-
-pub fn cloud_as_yew_img(b64: &str) -> yew::Html {
-    html! {
-        <div class="cloud">
-            <img alt="wordcloud" src={format!("data:image/png;base64,{b64}")} />
-        </div>
     }
 }